requirements.txt
# OpenAI Python client library for OpenAI-compatible endpoints
openai>=1.0.0

# Optional: For better environment variable handling
python-dotenv>=1.0.0

# PDF processing libraries
PyMuPDF>=1.23.0
camelot-py[cv]>=0.10.1
tabula-py>=2.5.1
pdfplumber>=0.9.0
pandas>=1.5.0
openpyxl>=3.0.0

# Image processing for PDF to JPEG conversion
Pillow>=9.0.0